# docs/openapi.json
{
  "components": {
    "schemas": {
      "Actor": {
        "properties": {
          "manager": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "enum": [
              "MANAGER",
              "EMPLOYEE"
            ],
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CategoryResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "unitId": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "CorrectDisposalDestinationRequest": {
        "properties": {
          "destination": {
            "enum": [
              "RECYCLING",
              "LANDFILL",
              "DONATION"
            ],
            "type": "string"
          }
        },
        "required": [
          "destination"
        ],
        "type": "object"
      },
...
      "CreateDisposalRequest": {
        "properties": {
          "destination": {
            "enum": [
              "RECYCLING",
              "LANDFILL",
              "DONATION"
            ],
            "type": "string"
          },
          "disposedAt": {
            "format": "date",
            "type": "string"
          },
          "itemIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "minItems": 1,
            "type": "array"
          },
          "notes": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "placeId": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "placeName": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "destination",
          "itemIds"
        ],
        "type": "object"
      },
...
      "DisposalReport": {
        "properties": {
          "issued_at": {
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Line"
            },
            "type": "array"
          },
          "owner": {
            "type": "string"
          },
          "proposal_deadline": {
            "type": "string"
          },
          "quote_number": {
            "type": "string"
          },
          "requester": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "DisposalResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "format": "uuid",
            "type": "string"
          },
          "createdByName": {
            "type": "string"
          },
          "destination": {
            "enum": [
              "RECYCLING",
              "LANDFILL",
              "DONATION"
            ],
            "type": "string"
          },
          "disposedAt": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DisposedItemResponse"
            },
            "type": "array"
          },
          "notes": {
            "type": "string"
          },
          "placeId": {
            "type": "string"
          },
          "placeName": {
            "type": "string"
          },
          "totalWeightKg": {
            "format": "double",
            "type": "number"
          },
          "unitId": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "DisposedItemResponse": {
        "properties": {
          "displayCode": {
            "type": "string"
          },
          "itemId": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "weightKg": {
            "format": "double",
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "ItemResponse": {
        "properties": {
          "acquiredAt": {
            "format": "date",
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "condition": {
            "enum": [
              "NEW",
              "USED",
              "SEMI_DAMAGED",
              "DAMAGED"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "format": "uuid",
            "type": "string"
          },
          "createdByName": {
            "type": "string"
          },
          "damages": {
            "items": {
              "enum": [
                "BROKEN_SCREEN",
                "MISSING_PART",
                "DOES_NOT_POWER_ON",
                "OXIDATION",
                "OTHER"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "displayCode": {
            "type": "string"
          },
          "hasDamages": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "lastEventAt": {
            "format": "date-time",
            "type": "string"
          },
          "manufacturingYear": {
            "format": "int32",
            "type": "integer"
          },
          "missingFields": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "model": {
            "$ref": "#/components/schemas/ModelResponse"
          },
          "name": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "predictedFailureDate": {
            "format": "date",
            "type": "string"
          },
          "predictionUpdatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "serialNumber": {
            "type": "string"
          },
          "status": {
            "enum": [
              "DRAFT",
              "PENDING_APPROVAL",
              "REJECTED",
              "IN_STOCK",
              "IN_MAINTENANCE",
              "AWAITING_EVALUATION",
              "DISPOSED",
              "REMOVED"
            ],
            "type": "string"
          },
          "unitId": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "usageIntensity": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "Line": {
        "properties": {
          "asset_number": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "equipment_type": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "origin": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "serial_number": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ModelResponse": {
        "properties": {
          "approvalStatus": {
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ],
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryResponse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "format": "uuid",
            "type": "string"
          },
          "estimatedWeightKg": {
            "format": "double",
            "type": "number"
          },
          "expectedLifespanMonths": {
            "format": "int32",
            "type": "integer"
          },
          "hazardous": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialResponse"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          },
          "reviewedAt": {
            "format": "date-time",
            "type": "string"
          },
          "reviewedBy": {
            "format": "uuid",
            "type": "string"
          },
          "unitId": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "warrantyMonths": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "PageResponseDisposalResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/DisposalResponse"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UpdateItemRequest": {
        "properties": {
          "acquiredAt": {
            "format": "date",
            "type": "string"
          },
          "condition": {
            "enum": [
              "NEW",
              "USED",
              "SEMI_DAMAGED",
              "DAMAGED"
            ],
            "type": "string"
          },
          "damages": {
            "items": {
              "enum": [
                "BROKEN_SCREEN",
                "MISSING_PART",
                "DOES_NOT_POWER_ON",
                "OXIDATION",
                "OTHER"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "hasDamages": {
            "type": "boolean"
          },
          "manufacturingYear": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "notes": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "serialNumber": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "usageIntensity": {
            "format": "int32",
            "maximum": 10,
            "minimum": 0,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API de gestão de inventário: categorias, modelos, materiais, itens (com ciclo de vida e manutenção), descartes, regras e dashboard.",
    "title": "ms-inventory",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/categories/{id}": {
      "delete": {
        "operationId": "delete_3",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "category-controller"
        ]
      },
      "get": {
        "operationId": "findById_4",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "category-controller"
        ]
      }
    },
...
    "/api/v1/disposals": {
      "get": {
        "operationId": "findAll_3",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "destination",
            "required": false,
            "schema": {
              "enum": [
                "RECYCLING",
                "LANDFILL",
                "DONATION"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "disposedFrom",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "disposedTo",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "createdFrom",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "createdTo",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "createdBy",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "placeId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "itemId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDisposalResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "disposal-controller"
        ]
      },
      "post": {
        "operationId": "create_3",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDisposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DisposalResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "disposal-controller"
        ]
      }
    },
    "/api/v1/disposals/{id}": {
      "get": {
        "operationId": "findById_1",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DisposalResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "disposal-controller"
        ]
      },
      "patch": {
        "operationId": "correctDestination",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectDisposalDestinationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DisposalResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "disposal-controller"
        ]
      }
    },
    "/api/v1/disposals/{id}/cancel": {
      "patch": {
        "operationId": "cancel",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "disposal-controller"
        ]
      }
    },
    "/api/v1/disposals/{id}/report": {
      "get": {
        "operationId": "report",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DisposalReport"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "disposal-controller"
        ]
      }
    },
...
    "/api/v1/items/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "item-controller"
        ]
      },
      "get": {
        "operationId": "findById",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "item-controller"
        ]
      },
      "patch": {
        "operationId": "update_1",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "item-controller"
        ]
      }
    },
...
    "/api/v1/models/{id}": {
      "delete": {
        "operationId": "delete_2",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "model-controller"
        ]
      },
      "get": {
        "operationId": "findById_3",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "model-controller"
        ]
      }
    },
...
    "/api/v1/rules/{id}": {
      "delete": {
        "operationId": "delete_1",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "rule-controller"
        ]
      },
      "get": {
        "operationId": "findById_2",
        "parameters": [
          {
            "in": "header",
            "name": "X-Unit-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "rule-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}